io: use correct register offset depending on the PCI bridge type
Depending on the type of a PCI bridge the layout of the PCI config space
is different. Currently io does not consider this when detecting the PCI
capabilities. On a computer with a PCI cardbus bridge this results in an
endless loop.

This change passes the header type as a parameter to the constructor of
Pci::Pci_pci_bridge_basic and Pci::Dev which eventually initializes the
`hdr_type` public member of Pci::Dev. The find_pci_cap() function uses
the value of `hdr_type` to determine the correct register offset.

Change-Id: Ic8dd112be9de2858218263615c76dba03745caf5