io: change d_printf into normal printf
Using d_printf while still parsing the cmdline with arg_init() will not
result in any output because the verboseness level is only set after
arg_init() is finished.

This change changes the d_printf used in the transparent MSI option into
a normal printf so that the user can actually see the output.

Change-Id: I1d8d7d37b1243cddabb37173387e55185622a9f5