Fix device resource allocation dilemma
We need to make sure that we handle a not yet allocated resource +
device hierarchy, for example, an uninitialized PCI bus correctly.
Therefore we add to extra member functions to the Resource_space API
that allow to first assign child resources to a not yet allocated parent
resource and then later adjust the whole resource hierarchy when the
topmost resource is allocated.

Change-Id: I3bca3b016171e5dae8a5d498d0508837820849cb