io: Add DMA domain for ARM IOMMUs
This change introduces the class Arm_dma_device, a device extension
with specific properties for use with ARM IOMMUs. Currently there is
support for SMMU-400 and IPMMU compatible IOMMUs.
For a detailed description of the ARM SMMU-400 support, see design ticket #737
For a detailed description of the IPMMU support, see design ticket #770

Change-Id: Iff58e0330a51ff07048d8a356ce1a02fa35eea29
Requires-Change-Id: I96357bc03ea889c0e72bf76c13ceca4935de6436