In Sw_irq_pin::_unbind() don't remove the IRQ type
Even without that fix I don't see any misbehavior, perhaps because the
IRQ type is currently not that important.

The IRQ type is part of the ''Sw_irq_pin::_state'' field and set in the
''Sw_irq_pin'' constructor. We must not clear this information because
it's never set again. I found this in the following scenario:

 1. IRQ pin (type `Irq_type_level_high`) attached to client 1
 2. same IRQ pin attached to client 2 which allocates Irq_mux
 3. client 2 terminates, process vanishes
 4. client 2 is restarted and the same IRQ pin is again attached to this
    client. The IRQ type (part of _state) is now 0 and remains at this
    value.

In 4 we enter ''Sw_irq_pin::bind()'' which calls ''Sw_irq_pin::bound()''
first to check if the IRQ pin is still bound to a client. Within that
function we do ''_irq.validate()'' but ''l4_error(_irq.validate())''
returns != 1 so we assume that the IRQ is deleted and therefore call
''_unbind(true).

In ''Sw_irq_pin::_unbind()'' we finally clear the ''_state'' field which
is IMO wrong because we also remove the `Irq_type_level_high` bits which
were originally set in the ''Sw_irq_pin'' constructor: The flags passed
to the constructor were 0x300001.

Change-Id: I5f7afb9a01d1353ccea13f8210892a2a75dcbe3a