Fix bogus documentation of out parameter
The device() method and the l4vbus_get_device function return
information about a device. They expect a pointer to a structure where
the device information shall be put after a successful call.

This parameter is optional and if NULL is passed no information is
returned. The documentation for the devinfo parameter suggested that
the pointer may be set to NULL by the called function which is wrong.
This corrects the documentation.

Change-Id: I271a67edf0d2962762db6e30d7b5a092f34a0c11