io: extend op_msi_info() to work with device handles
To get the MSI info for an IRQ the client has to provide the full source
ID which may be cumbersome to aquire (may involve a full PCI bus scan).

Thus we allow the usage of a device handle as the source ID as well. The
caller must set the L4vbus::Icu::Src_dev_handle flag to denote that it
intends to use the device handle.

Change-Id: I2552db7a619ad333153e38796bdaf260376c073d