io: i915 driver: fix allocation of OpRegion
The IGD OpRegion is a mechanism for the OS and system firmware to
collaborate on various functionality. For example it is used for
communication between the i915 graphics driver and ACPI. The BIOS places
the memory offset of the OpRegion into the ASLS PCI register. The
default value of this register is 0h.

This change adds a sanity check whether BIOS actually has set a memory
address to avoid giving the user access to physical memory at address
0h. The OpRegion is only added as a resource if the ASLS register is not
0.

Thanks to Christian Grönke!

Change-Id: I63713905668a745bb2c990b2ffd8bb39d893c219