io: fix Iostream handling of GPIO Pin_slice
In the Gpio class multi* functions, extracting a received Pin_slice
object using the Istream::get() method imposes an alignment on the
object that does not always match that used when the Pin_slice's offset
and mask components are added individually with Ostream::put().

The Pin_slice object is used in the Gpio class implementation and the
libvbus C++ API but it is added to the Iostream as unsigned offset and
unsigned mask using the libvbus C API.

Converted to extract the Pin_slice members individually as two unsigned
integers.

Change-Id: I7782678e7f9a5472893cc0d9ee790872d2b5eb16