io: fix 'client' in request_child_resource() call
The Pci_pci_bridge_irq_router_rs resource space is used as IRQ router of
a PCI bus if the corresponding PCI bridge does not define its own IRQ
router. The Pci_pci_bridge_irq_router_rs allocates the interrupt
resource from the parent. Going up the bus chain we eventually reach the
root but with the Root_irq_rs resource space. The request method of
Root_irq_rs ignores the cld parameter and therefore this change will not
change the behavior with for setups. Nevertheless we should pass the
correct parameter.

Change-Id: I930eb446fa2986db0ce37b542e22e71802dfcb32