io: fix assignment of prefetchable client resources
Like in Generic_device::request_child_resource() we add a second run to
Generic_device::alloc_child_resource() which allows to assign a client
resource of prefetchable memory to a parent resource of non-prefetchable
memory.

This case is required if, for instance, a PCI root bridge provides only
a non-prefetchable MMIO resource but no prefetchable MMIO resource.

Change-Id: I159e14faca6d0a5a5bc4800310f203e09e427b8b