move Resource_provider class into its own file
Resource_provider is only used for PCI and ACPI. Removing it from
the generic resource header means that it is also no longer
exported into the Lua interface and thus can be optimised
away by the linker when PCI and ACPI are not compiled in.

The Lua interface never allowed instantiation of Resource_provider
or provided access to any of its functions. Also none of the
exported function takes Resource_provider as a parameter or
returns it. Therefore we don't need the conversion functions
that are still exported.

Change-Id: I0998b65ce544fecf5abb98ac5cf42876eeed8674