io: print error if PCI ACS config failed
Certain Intel PCIe root ports do not implement the PCI ACS capability
according to the PCI Sig specification. Instead they use a dword for
both the capability and the control register.

To avoid giving the user a false impression of a properly configured
system we read the control register back and compare it with the desired
value. This assumes that writing into the 32bit read-only capability
register on broken implementations actually does not modify the upper
16 bits.

See ticket #CD-155.

Change-Id: I246fc9fefceebc4c6e6fc5642feb43f4618c90b3