io: use local path for pciids.h
This is required for unit tests: Several files in io/server/src include
the global visible '<l4/io/pciids.h>' thus we cannot just fetch these
files with PRIVATE_INCDIR.

In the IO unit test I would like to display the PCI IDs for debugging
reasons. The unit test fetches files from IO and adds mocks for the
missing components.

As libpciids is not a real package, it cannot be listed as dependency
for the IO test.

Change-Id: I4204b2b802dd160b3826e4551d555d340974fdf0