io: don't include PCI ID header if disabled
Only include "pciids.h" if enabled by configuration. This worked before
because the headers are always available and the corresponding functions
are only used if using the PCI IDs is enabled by configuration. But with
the current change it is possible to compile the code if the PCI ID DB
headers don't exist.

Change-Id: I26bdfd5efbfc8b100263dea1a53e003aeaded0ba