io: fix PCIe extended capability discovery on qemu
When io finds a PCIe capability for a PCI device it starts scanning for
the extended capabilities. The absence of extended capabilities is
indicated by an extended capability header of 0x0000_0000. io uses that
condition to break the scan loop.

Qemu allows to add a PCIe device to an emulated PCI bus and the PCIe
device still exposes the PCIe capability. However, reading the first
extended capability from the extended config space only returns
0xffff_ffff. This catches io in an endless loop. See ticket #CD-249 for
more details.

This patch adds another check to make sure an extended capbility is
valid. This makes io break the scan loop.

This fixes ticket #CD-249.

Change-Id: Ibbef617023c1eb3f50b8d1cf4a6883333d15d6b5