io: zero PCI config space of virtual PCI devices
The Pci_virtual_dev's ctor contained a memset zeroing a pointer
variable. The variable is set in the derived classes, hence, implying
to zero the area pointed to.
Ctor initialization order is base classes first, most derived last,
leading to a false conclusion.

This commit removes the memset and the then empty ctor and adds it to
the ctors of the derived classes.
It also defaults the dtor of Pci_virtual_dev.

Change-Id: I300989b02f1b6bd18dd7cd4fbb2120462ff64024