io: use request/alloc terminology for Phys_space
In principle, Phys_space manages a resource space. The interface
Resource_space uses request for regions with defined start and end
address and alloc for regions which have a size and alignment
constraint but no defined address.

To use the same name for the same thing, the overload of alloc
expecting a Phys_space::Phys_region parameter is renamed to request.

The only caller of this now correctly named function is
Root_mmio_rs::request(), which is derived from  Resource_space.

Change-Id: I6997892a12461a0efb1e2b5e3dc841e054fb3bc0