io: add_child() should check for parent before modifying depth
If add_child() gets a device that already has a parent, then the
function first modifies the depth of the device and then refuses
to add the device in the tree. This leaves the child potentially
in an inconsistent state within the tree it was already in.

This patch changes the order and first checks that the device has
no parent and afterwards ajusts the depth.

Change-Id: Idb51dd33486f62bcb3abf8e398d170c97edb07fc