io: remove unnecessary virtual functions in resource allocation
Generic_device provides the only implementation for the
virtual functions request/alloc_child_resource(). All devices
inherit from that class and use it. All functions used in the
implementation are already available in the Device class.
So there is really no need to have a virtual function here.
Move the implementation directly into Device instead.

This also makes the code a bit more consistent as now all
functions related to resource allocation are implemented
in the same class.

Change-Id: I786a0d9dc71020038f417aa7abfc7460b5f4bc94