io: Add register specific PCI read and write
This change adds read and write functions to the Pci::Cfg_rw_mixin
that allow config-space accesses with special per-register data types.
Such a datatype must provide a data member in the correct size (fitting
the config-space register) and a static 'Ofs' member that specifies the
offset of that register in bytes.

This feature allows us to provide size and offset properties for PCI
config-space accesses in a single encapsulated data type that might also
include additional bitfield definitions.

Change-Id: I1d67116c2179fda158fb401e4bc40992babed9fd