vbus: fix formatting issues in documentation
In the generated html documentation for l4vbus_get_device_by_hid,
the `parent' parameter description is interlaced with the function
description. The proper solution to this issue would require a
command like \copyparams, which doxygen currently does not implement.
The next best thing is to add a line break to separate the parameter
description from the function description.

Change-Id: Ib11b8deb59a982341eef4d6e6816884e95eeed9f