Avoid type narrowing error by using a type
Give the vbus device ops a type to avoid the type narrowing issue in
the switch-case statement:

  error: narrowing conversion of ‘L4vbus_vdevice_hid’
         from ‘int’ to ‘unsigned int’ [-Wnarrowing]

  case L4vbus_vdevice_hid:
       ^~~~~~~~~~~~~~~~~~~

Reason for this error:

  L4VBUS_INTERFACE_GENERIC << L4VBUS_IFACE_SHIFT = 0x80000000

A proper fix would be to set the type of 'enum L4vbus_vdevice_op' to
'l4_uint32_t'. However, that's a C++-only feature.

We are allowed to cast 'func' to 'L4vbus_vdevice_op' in the switch
statement because the previous line already ensures that we only need
to care about L4VBUS_INTERFACE_GENERIC-based constants.

Change-Id: I86b53dd2b88e9aadd8065dfc61c02cb82b87c0d0