io: prevent infinite loop while requesting IO ports
If a caller requests an IO port range where the start of the range is
larger than the end, io ends up in an infinite loop calculating the
flexpage size. The error is caused by an unsigned overflow.

This change adds detection for the overflow and returns with an error in
such cases.

This fixes #CD-324.

Thanks to Stefan Fritsch for reporting this issue.

Change-Id: I352139e51712686ac049d9bca08ca23b54714cf0