io: scan PCI options ROMs and add them as resource
Currently io does not touch the expansion ROM configuration of a PCI
device. Expansion ROMs are also not passed on to clients.

Because io does not account for expansion ROMs they are not treated as a
device resource. It might now happen that firmware setups the whole PCIe
bus but disables the MMIO decoders of certain devices. Io treats those
MMIO BARs as not allocated. If now the addresses programmed by firmware
are in the following order: Option ROM < BAR1 < BAR2, io will move the
addresses of BAR1 and BAR2 during the allocation step. Now the address
of BAR1 might alias with the Option ROM BAR. According to the PCI spec
this shouldn't be a problem but in reality it is (looking at you Intel
i210 device).

Thus I changed the handling and treatment of the Option ROM. Just to be
sure all Option ROMs are now disabled by io. I also add the Option ROM
as a resource to the physical device in order to make the resource
allocation work. As previously the Option ROM is not passed on to
clients. That has to be explicitly enabled in io by changing a config
option.

Change-Id: I70b84a91ce1ef17f636026c9d20197123bbb9c2f