Fix provider_device_handle in vgpio
In 67e808 the handle code changed from pointers to indices. The case an
resource uses a different provider was not updated.

We can't set the handle number directly, because at the time the handle
is still unknown. Save the provider itself and return the handle at the
time when it is needed.

Change-Id: I23062e19f896c51c98906440173642edebad2780