io: add 4GB check for MMIO regions on 32-bit systems
On 32-bit systems, IO can handle physical addresses up to 4GB. If an
MMIO resource occupied a region beyond 4GB on a 32-bit system, the
address was silently cropped and the region of the cropped address
was mapped.

Fix that problem and add a proper check in res_map_iomem() so that the
user knows what's going on.

Change-Id: I2ed18d8e31f9e835d01e74cbff01ba5bea85a858