io: change resource request for PCI device resources
Currently it depends on the status (enabled / disabled) of the address
decoders how MMIO and IO port resources of a PCI device are requested.
The code seems to assume that the status of the decoder is an indicator
for the validity of the stored resources.

However this leads to io making false assumptions and later reallocating
(and moving) resources. This may cause problems with e.g. L4Linux who
then get resource conflicts and again reprogram the device.

This change removes the F_disabled flag from the resource flags.

Change-Id: Ifc35e40e3ba9b26c1e61004dbcbf46bf276b1829