io: add ACPI support for arm64
Compile most of the ACPI code for arm64 too. Because Arm does not have
port-IO, the Root_brige must not be used in any way.

All information about the wiring of PCI root complexes, SMMUs and ITSs
(Arm MSI controller) is found in the Arm specific IORT ACPI table. See
ARM DEN 0049E for the details.

To map the ITS for the device, io creates mappings starting from
0xf0000000 in the DMA task. Because we lack a consistent DMA address
space management, this information is not known to moe or uvmm. In case
other mappings are created in this range, MSIs probably won't work. See
ticket #CD-171.

Also, if any other device PCI BAR is in that range, MSIs may not work
either.

Change-Id: Id59b21f0e8b801613889987ce82fc290c04a9e18