Replace read-only rights flag with separate read and write flags
When creating MMIO resources the read and write rights must now be
specified. For all existing code locations RW rights have been added,
with the exception of ACPI memory resources, where the ACPI WriteProtect
property is now respected.

To maintain backwards compatibility with existing io configurations,
the io lua wrapper has been adapted, so that if access rights are not
explicitly set, RW is used as default.

A detail worth mentioning is that memory which has the read flag is
mapped not only with read rights but also execute rights are included.
The rationale being that Io is not responsible for execute rights, in
general in almost any case the application that maps memory is under
control of execute rights The execution right exists for good reasons,
because MMIO areas can be ROMs from a device that may contain code, or
something similar.

Change-Id: I043e978fe807062bda83afd3e77643e03a0158e3