Generalize ARCH-ABI support for common make variables
It was always possible to write something like:

 SRC_CC_arm-l4f += arm-test.c

However, using opts like in

 SRC_CC_amd64-l4f-$(CONFIG_FEATURE) += amd64-feature.c

wasn't supported for all variables. Generalize this for the following
vars:

 CFLAGS CXXFLAGS CPPFLAGS SRC_C SRC_CC SRC_S DEFINES LD_FLAGS
 REQUIRES_LIBS

Please note that it is required to specify ARCH + ABI. For L4Re
applications this is ARCH-l4f. Plain apps are just using ARCH.

Fix all packages to use the correct notation.

Change-Id: Ia203ff241806e60240bd8d9349f986d68e84276e