io: remove ARI-style device discovery
Currently io unconditionally enables ARI forwarding on all PCIe
downstream ports to discover ARI devices. According to the PCIe spec ARI
forwarding should only be enabled in a downstream port if the device
immediately below the downstream port is an ARI device (see section 6.13
Alternative Routing-ID Interpretation (ARI)). See the respective
implementation note for more information.

Until now this wasn't a problem. A customer now reports that system
firmware (Coreboot) is unable to power down a machine due to being stuck
on such a ARI-forwarding-enabled downstream port. This seems to be
caused by a bug in the system firmware itself but io should adhere to
the standards as well.

As ARI forwarding is not needed right now this change removes the code
that enables that feature on a downstream port. Please be aware that
this also breaks the support for SR-IOV devices.

Change-Id: I9092008b5de37539a450936d182517ed58b57c3b