io: add Generic_device::assert_property()
This function can be used to verify that a Property has not the default
value -- it was set in the Lua script.

This function simplifies the device setup code: The function prints a
proper warning (if the ERR debug level is enabled) using the device
name. The availability of the device name available is the reason for
not making this function a member of 'Property'.

Change-Id: I91f78f127471c7f97df7fe94fcbd93184189522d