Restore rights for mmio dataspaces
In change I043e978fe807062bda83afd3e77643e03a0158e3 the way how rights
are handled was changed to require the rights to be made explicit
instead of a read-only flag being used to restrict them and a default to
all rights.

This change caused MMIO Dataspaces to not be mappable to clients since
their constructor did not specify any rights on resource construction.

Specify both read and write rights for these types to restore the
previous behavior

Change-Id: Iaf3d2d5083cfe5426e01ce0439d15c3980d01e89