io: upgrade ACS enabled message
In order to validate that PCI ACS has been enabled it would be nice if
io always prints a message. The current success message is hidden behind
a debug level, which usually is not active in a production system. Such
a message would also make (automatic) testing whether the feature has
been enabled a lot easier.

Thus we change the message to use a normal printf such that it gets
printed always.

Change-Id: Ib96a85bc65afdb6ecb69a36a119a503784df7c2e