acpi: disable warning at the correct place
The BID mk rules append -Wall to CPPFLAGS basically reverting the effect
of disabling a warning. It is also not sufficient to append to CPPFLAGS
after including the mk rules because $(WARNINGS) is appended to CXXFLAGS
after $(CPPFLAGS). Hence, append the parameter for disabling the warning
to WARNINGS and do this after evaluating the mk rules.

Change-Id: Ibbe0d98025cbc4e35620b3c530cfd65f3215c62e