io: Arm_dma_domain: Adjust to supported IOMMUs in Fiasco
The ARM IOMMU support in IO was added long before ARM IOMMU support
landed in Fiasco. It turns out the current code in IO is actually
completely broken because it does no longer match the interface in
Fiasco:

  - The dev_id (now stream_id) was extended to 32 bit and the iommu_idx
    (now smmu_idx) shifted to bits 32-47.
  - Support for the Renesas IPMMU never landed in Fiasco.

Right now Fiasco supports only standard ARM SMMUs. Confusingly, the
current code for SMMU-400 expects a "sidreg" property that points to
a register space where the stream IDs can be dynamically allocated
to a device. This functionality is actually not part of a standard
SMMU but is completely separate and specific to the Layerscape LS1021
platform. It would be much more clear to handle that in a separate
driver in IO or even allocating the stream IDs statically in e.g.
bootstrap already.

Make the code work again for now by replacing all the existing
properties with a simple "sid" property that allows assigning
a fixed SMMU stream ID to a device in the IO configuration.

Change-Id: I7e7652c83e25e9fcafdafc1314e5b7c02181306e