acpi: Evaluate HID descriptor address for I2C devices
The HID descriptor address of HID devices on an I2C bus is provided by
ACPI via an Device Specific Method (DSM) as specified in the HID over
I2C specification.

Du to Io's limited resource description structure the HID descriptor
address currently cannot be stored as part of the device, therefore
it is printed as debug info instead.

Change-Id: I0f7d57e7dbf2a8804011751a055a87eb92bd30be