acpi: Add GPIO resources
Io scans device information provided via ACPI. For the devices
discovered in the process, Io creates a device object each and adds
resource descriptions according to the resources specified in ACPI.

This change adds support for GPIO resources that have not been
considered so far.

Of the GPIO resource information provided via ACPI, currently only
the PIN number and whether the pin is used as an interrupt, can be
represented in Io's resource description.
Additional information that cannot be represented due to Io's limited
resource description structure is printed as debug info.

Change-Id: I469c48a13d790dc334c0fd5499df4de909b7704b