io: move MSI source id translation to platform code
The MSI source id that is passed to Icu::msi_info() is platform
dependent. The PCI code should not handle the details of the MSI-ID at
all but expects a platform adapter object to take care of that. The
Platform_adapter_if object has to deal with the platform specific
details and is injected as dependency when the root bridge is created.
Right now we only support the Intel VT-d format on x86 platforms.

The Vtd_platform_adapter additionally has to take care of the PCI
bridge aliasing that can arise. The translate_msi_src() method has to
walk the bus hierarchy upwards to take care of source-id rewriting of
certain bridges.

No change in observable behaviour is intended.

Change-Id: I9e8a8011bc4c8856761fc1b5571a9ef37cea6310