io: remove unused Msi_mgr
The list of Msi_mgr instances is not used anywhere. There is neither
code that walks the list nor any caller for Msi_mgr::reconfig_msi().
Also, the Hw::Pci::Dev::_msi_mgrs list does not own the objects so
removing the list does not change the lifetime of the objects either.

Change-Id: I68a85439bf17a03dc7bd00ad43a65f2f9a301179