io: add infrastructure to map MSI controller
On Arm the PCI device writes are not handled special in any way by the
IOMMU. It is the responsibility of the operating system to install a
mapping in the IOMMU page table of the device so that the respective
MSI controller can be reached.

Here we just add the infrastructure for the platform code to install
these mappings. On x86 nothing needs to be done.

Change-Id: I043e011f5c82cb5bf3c8ce1b50ddc5b179770520