io: move DMA source id translation to platform code
The DMA source ID that is needed in Iommu::bind() is platform
dependent. The PCI code should not handle the details of how the ID is
structured but rely on the PCI platform adapter to take care of it.

Because some bridges take ownership of device writes, we must bind the
device with each possible alias to the IOMMU. The old code relied on
some Fiasco extension to bind to every device on a bus in this case.
This is not necessary and we explicitly bind to every possible alias
explicitly.

No change in observable behaviour is intended.

Change-Id: I917342262351030656a8c5e338c6c6cc46ceb940