use unsigned to represent resource indexes
In order to adhere to Misra rules M5-0-4 and M5-0-5, which
forbid implicit integral conversions that change signedness,
the l4vbus interface must use unsigned, or the clients must
be adapted to call the interface providing an index of type
int. Changing the signature of the affected operations to
unsigned leads to a clearer semantics for the index parameter.

See ticket #CD-1105.

Change-Id: I598ae4aee071d9aae80d0a415ba56cdaf946f78f