io/pci: Add SR-IOV support
Add support Single Root I/O Virtualization and Sharing (SR-IOV), where a
single device can provide multiple "light-weight" PCIe Functions,
referred to as Virtual Function (VF), which are managed by a Physical
Function (PF).

The approach taken here is for io to perform generic PCI SR-IOV
enablement of SR-IOV capable device. Both the PFs and the VFs are
represented as PCI devices on the vbus. Then a "manager" vbus client
takes care of the device-specific configuration of the PF. As the VFs
are isolated from each other, they can be passed on to arbitrary
clients, for example to a VM guest.

Current limitations:
  - SR-IOV is always enabled if the device supports it.
  - Number of VFs to enable is configured statically according to the
    L4IO_PCI_SRIOV_MAX_VFS config option (limited by bus capacity and
    number of VFs the device supports). In the future configuring the
    number of VFs might be made dynamic, however that would require
    preparatory work to make vbus more dynamic.
  - Allocating multiple bus numbers for a SR-IOV device not supported.
  - Dependencies between PFs are not supported.

Co-authored-by: Alexander Warg <alexander.warg@kernkonzept.com>
Change-Id: I53d69e907f1206ef8f04448f1115da749799fbb8