Control: Make libpciids explicit
Io provides the libpciids (the PC file is generated and used internally)
but does not advertise it. While this works, as an internal library, it
will impede checking externally that requires and provides statements
are consistent with the REQUIRES_LIBS list.

Since libpciids is in REQUIRES_LIBS for io, automated checks will
determine that it is neither provided by the package itself, nor is it
required from another package.

To make automated checking possible we should advertise that this is
actually created.

Change-Id: Ieab58faec9fdb02350343bbc6dc6ff74f8a9dc42