io/x86: fix device to IOMMU matching
When two IOMMUs overlap with regards to the start bus number and end
bus number, a device which is not explicitly mentioned in the DMAR
is matched to the first IOMMU in the list.
This behavior is wrong, when the IOMMU does not manage a PCI
sub-hierarchy and when it does not act as default IOMMU.

This commit disentangles the code paths and stores the information from
the Device Scope entry of a DRHD entry. This is then matched for the bus
number instead and only relies on the referenced IOMMU to match the
segment field.

It can now also handle multiple subhierarchies behind an IOMMU.

Change-Id: I1b5ffe8681761d0b7cd8667ccde5f3c31e48dc13