io: Fix strict aliasing violation
The compiler has to assume that `ACPI_TABLE_HEADER` can alias
`ACPI_TABLE_ECDT`, because `ACPI_TABLE_ECDT` contains a
`ACPI_TABLE_HEADER Header;` member but the same does not apply to
pointers. Therefore, the compiler might assume that `ecdt` will not be
modified by `AcpiGetTable()` because a `ACPI_TABLE_HEADER **`, as passed
to the function, cannot alias with `ACPI_TABLE_IORT **`.

Change-Id: I30033ce87438fc0bec518ce374e2208baa84dbcd