[API] io/libvbus: introduce L4VBUS_RESOURCE_F_MEM_{CACHE,PREFETCH}ABLE
These flags are specific to IO memory regions.

If L4VBUS_RESOURCE_F_MEM_CACHEABLE is set, the client may map the region
cached or non-cached. If this flag is not set, the client _must_ not map
the region cached.

If L4VBUS_RESOURCE_F_MEM_PREFETCHABLE is set, the client may map the
region buffered or non-cached.

See also ticket #CD-1426.

Change-Id: I2cbe7edaa02c82c481e662f628770f931ce0b0f4