io: rename iMX8MP GPIO devices and instantiate driver
By creating the right device type, io is capable of driving the GPIO
controllers of the iMX8MP. Also rename the devices to align with their
names in the Linux device tree which makes matching pins a lot easier.

This change also changes the size of the MMIO region for the GPIO
controllers to 0x1000. The actual registers only need 32 bytes.

Change-Id: Iedcfe8cc89cc948b796d4da7dc3d3cbb9f1dca4d