io: fix enumeration of secondary bridges
A bridge may have other bridges connected behind it. In this case, the
subordinate bus number will be greater than the secondary bus number.
In case of ACPI, the bus range is usually described by a bus number
range resource. Make sure we use the whole range and not just the first
secondary bus.

Change-Id: I96c3ddb783c981402399d408e99b5798c62e2ed7