Add --enable-sriov argument to select specific SR-IOV devices
As of now, CONFIG_L4IO_PCI_SRIOV enables the SR-IOV capability for
every device found on the bus. This breaks drivers of certain
devices and causes io to deadlock when it encounters certain
devices. To avoid such issues, this commit requires the user to
specify a "--enable-sriov $vendor_id:$device_id" argument.

Change-Id: Ie80bcb87155a3435531e3c5f82b448bfafe4f655
Signed-off-by: Benedikt Radtke <benediktradtke@gmail.com>